# Import/Perf: set fcurve interpolation using foreach_set