Closes #1174

@julienduroure's link there showed how to foreach_set an enum property.

|Model|Before|After|
|:---|:---|:---|
|Brainstem|0.9s|0.5s|
|Level.gltf (iss862)|9s|6s|
|VC|2.4s|0.5s|